+++ _examples/axon-application/src/main/avro/BankAccountContext.avpr
